Create a page for each recipe
I created a folder called "img" to store the images used on the recipe
pages and another folder called "pages" to store the html pages of the
recipes. The basic HTML code has been created for each recipe page and
includes an image of the recipe, a description, the ingredients and the
steps.